map subgraph functions with context same as rest of query library
after implementing downstream, it has become apparent that resolving and passing the `subgraphUrl` every
time is not a very good pattern. As the subgraph data must be injected into the context for the queries library anyway, this will
allow for calling of subgraph functions directly from `useSynthetixQueries` with the first `subgraphUrl` argument
mapped with the url of the corresponding subgraph.